View presentation refactoring
