+++ MvvmKit.Abstractions/Presenters/IAttributeViewPresenterHelper.cs
@@ -1,0 +1,8 @@
+using MvvmKit.Abstractions.View;
+
+namespace MvvmKit.Abstractions.Presenters;
+
+public interface IAttributeViewPresenterHelper
+{
+	BasePresentationAttribute GetPresentationAttribute(ViewModelRequest request);
+}
