--- MvvmKit.Abstractions/Presenters/IOverridePresentation.cs
@@ -1,8 +1,0 @@
-using MvvmKit.Abstractions.View;
-
-namespace MvvmKit.Abstractions.Presenters;
-
-public interface IOverridePresentation
-{
-	BasePresentationAttribute PresentationAttribute(ViewModelRequest request);
-}
