+++ MvvmKit.Abstractions/View/IViewPresenterResolver.cs
@@ -1,0 +1,8 @@
+using MvvmKit.Abstractions.Presenters;
+
+namespace MvvmKit.Abstractions.View;
+
+public interface IViewPresenterResolver
+{
+	IViewPresenter Resolve(ViewModelRequest request);
+}
