+++ Platforms/Ios/MvvmKit.Platforms.Ios.Abstractions/Presenters/INavigationControllerViewPresenter.cs
@@ -1,0 +1,7 @@
+using MvvmKit.Abstractions.Presenters;
+
+namespace MvvmKit.Platforms.Ios.Abstractions.Presenters;
+
+public interface INavigationControllerViewPresenter : IViewPresenter, ICanCreateIosView
+{
+}
