+++ Platforms/Ios/MvvmKit.Platforms.Ios/Views/IosViewDispatcher.cs
@@ -1,21 +1,22 @@
 using MvvmKit.Abstractions.View;
-using MvvmKit.Platforms.Ios.Abstractions.Presenters;
 using MvvmKit.Platforms.Ios.Threading;
 
 namespace MvvmKit.Platforms.Ios.Views;
 
 public class IosViewDispatcher : IosUIThreadDispatcher, IViewDispatcher
 {
-	private readonly IIosViewPresenter presenter;
+	private readonly IViewPresenterResolver viewPresenterResolver;
 
-	public IosViewDispatcher(IIosViewPresenter presenter)
+	public IosViewDispatcher(IViewPresenterResolver viewPresenterResolver)
 	{
-		this.presenter = presenter;
+		this.viewPresenterResolver = viewPresenterResolver;
 	}
 
 	public async Task<bool> ShowViewModelAsync(ViewModelRequest request)
 	{
-		Task Action() => this.presenter.ShowAsync(request);
+		var presenter = this.viewPresenterResolver.Resolve(request);
+
+		Task Action() => presenter.ShowAsync(request);
 
 		await ExecuteOnMainThreadAsync(Action);
